+++ crates/tui-realm-stdlib/examples/container.rs
@@ -145,7 +145,7 @@ impl Default for MyContainer {
                             .rewind(true)
                             .step(4)
                             .row_height(1)
-                            .headers(&["Key", "Msg", "Description"])
+                            .headers(["Key", "Msg", "Description"])
                             .column_spacing(3)
                             .widths(&[30, 20, 50])
                             .table(
@@ -194,7 +194,7 @@ impl Default for MyContainer {
                             .highlighted_color(Color::Green)
                             .highlighted_str(">> ")
                             .row_height(1)
-                            .headers(&["Key", "Msg", "Description"])
+                            .headers(["Key", "Msg", "Description"])
                             .column_spacing(3)
                             .widths(&[30, 20, 50])
                             .table(