+++ crates/tui-realm-stdlib/examples/paragraph.rs
@@ -148,7 +148,7 @@ impl Default for ParagraphAlfa {
                 .background(Color::Black)
                 .title("Lorem ipsum (wrap)", Alignment::Center)
                 .wrap(true)
-                .text(&[
+                .text([
                     TextSpan::new("Lorem ipsum dolor sit amet,").underlined().fg(Color::Green),
                     TextSpan::from("consectetur adipiscing elit. Praesent mauris est, vehicula et imperdiet sed, tincidunt sed est. Sed sed dui odio. Etiam nunc neque, sodales ut ex nec, tincidunt malesuada eros. Sed quis eros non felis sodales accumsan in ac risus"),
                     TextSpan::from("                       Duis augue diam, tempor vitae posuere et, tempus mattis ligula.")
@@ -185,7 +185,7 @@ impl Default for ParagraphBeta {
                 .background(Color::Black)
                 .title("Lorem ipsum (no wrap)", Alignment::Center)
                 .wrap(false)
-                .text(&[
+                .text([
                     TextSpan::new("Lorem ipsum dolor sit amet,").underlined().fg(Color::Green),
                     TextSpan::from("consectetur adipiscing elit. Praesent mauris est, vehicula et imperdiet sed, tincidunt sed est. Sed sed dui odio. Etiam nunc neque, sodales ut ex nec, tincidunt malesuada eros. Sed quis eros non felis sodales accumsan in ac risus"),
                     TextSpan::from("                                        Duis augue diam, tempor vitae posuere et, tempus mattis ligula.")