+++ crates/tui-realm-stdlib/examples/table.rs
@@ -153,7 +153,7 @@ impl Default for TableAlfa {
                 .rewind(true)
                 .step(4)
                 .row_height(1)
-                .headers(&["Key", "Msg", "Description"])
+                .headers(["Key", "Msg", "Description"])
                 .column_spacing(3)
                 .widths(&[30, 20, 50])
                 .table(
@@ -241,7 +241,7 @@ impl Default for TableBeta {
                 .highlighted_color(Color::Green)
                 .highlighted_str(">> ")
                 .row_height(1)
-                .headers(&["Key", "Msg", "Description"])
+                .headers(["Key", "Msg", "Description"])
                 .column_spacing(3)
                 .widths(&[30, 20, 50])
                 .table(