+++ crates/tui-realm-stdlib/CHANGELOG.md
@@ -27,6 +27,23 @@
 
 ---
 
+## next
+
+Unreleased
+
+- Change `Paragraph::text` to take a iterator to clone less values.
+- Change `Span::spans` to take a iterator to clone less values.
+- Change `Textarea::text_rows` to take a iterator to clone less values.
+- Change `Canvas::data` to take a iterator to clone less values.
+- Change `Chart::data` to take a iterator to clone less values.
+- Change `CheckboxStates::set_choices` to take a `Into<Vec<String>>` to clone less values.
+- Change `Checkbox::choices` to take a iterator to clone less values.
+- Change `RadioStates::set_choices` to take a `Into<Vec<String>>` to clone less values.
+- Change `RadioStates::choices` to take a iterator to clone less values.
+- Change `SelectStates::set_choices` to take a `Into<Vec<String>>` to clone less values.
+- Change `Select::choices` to take a iterator to clone less values.
+- Change `Table::headers` to take a iterator to clone less values.
+
 ## 2.0.1
 
 Released on 13/10/2024